requirements.txt
fastapi
uvicorn
python-multipart
opencv-python-headless
Pillow
ImageHash
httpx
python-dotenv
slowapi
sqlalchemy
passlib
bcrypt
PyJWT
